+++ SLK_img2pixel/changelog.md
@@ -1,10 +1,18 @@
 # Changelog
 
 
-Current development version: SLK_img2pixel 0.2
+Current development version: SLK_img2pixel 1.0
 
 ----------------------------------------
-Version:                SLK_img2pixel 0.2 - unicode hell (tba)
+Version:                SLK_img2pixel 1.0 - initial release (tba)
+----------------------------------------
+
+CHANGES:
+
+* preview now drawn by the gpu
+
+----------------------------------------
+Version:                SLK_img2pixel 0.2 - unicode hell (22.01.2021)
 ----------------------------------------
 
 CHANGES: